# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is tz api",
        "title": "Rest api",
        "termsOfService": "not much usefull",
        "contact": {
            "name": "Asliddin",
            "url": "https://t.me/asliddindeh",
            "email": "asliddinvstalim@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8070",
    "basePath": "/v1",
    "paths": {
        "/user": {
            "post": {
                "description": "this func create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "create user with info",
                "parameters": [
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.UserResponse"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "description": "this func get all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UsersResponse"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "this func get user info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "get user info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "user.UserRequest": {
            "type": "object",
            "properties": {
                "Age": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Phone": {
                    "type": "string"
                }
            }
        },
        "user.UserResponse": {
            "type": "object",
            "properties": {
                "Age": {
                    "type": "integer"
                },
                "Id": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Phone": {
                    "type": "string"
                }
            }
        },
        "user.UsersResponse": {
            "type": "object",
            "properties": {
                "Users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.UserResponse"
                    }
                }
            }
        }
    }
}